+++ index.js
@@ -8,7 +8,7 @@
  */
 
 export * as P from './src/promise.js';
-export * as PF from './src/point-free.js';
+export * as PF from './src/fantasyland.js';
 export * as M from './src/maybe.js';
 export * as E from './src/either.js';
 export * as T from './src/transformations.js';