+++ src/fantasyland.js
@@ -1,0 +1,5 @@
+export { default as ap } from './fantasyland/ap.js';
+export { default as chain } from './fantasyland/chain.js';
+export { default as liftA2 } from './fantasyland/liftA2.js';
+export { default as map } from './fantasyland/map.js';
+export { default as pluck } from './fantasyland/pluck.js';