+++ doc/changelog.md
@@ -2,7 +2,12 @@ Changelog v{{ config.meta.version }}
 =========
 0.2.0
 -----
-- `point-free` functions `chain`, `map`, `ap`, `liftA2` to support FantasyLand types *and* native Promise
+- renamed `point-free` submodule to `fantasyland`,
+- fixed tests for Node v16
+
+0.2.0
+-----
+- `point-free` functions `ap`, `chain`, `map`, `pluck`, `liftA2` to support FantasyLand types *and* native Promise
 - Fantasy-Land support for Cancelables
 
 0.1.43